make add feed a popup + minor edits
